вынесит общую логику обновления
